requirements.txt
anthropic>=0.49.0
httpx>=0.27.0
pydantic>=2.7.0
pyyaml>=6.0
